requirements.txt
accelerate==0.16.0
aiohttp==3.8.1
aiosignal==1.2.0
ansi2html==1.8.0
anyio==3.5.0
arrow==1.2.3
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
bootstrapped==0.0.2
boto3==1.24.91
botocore==1.27.91
brotlipy==0.7.0
bs4==0.0.1
cachetools==5.2.0
click==8.1.3
cycler==0.11.0
dash==2.13.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datasets==2.1.0
debugpy==1.6.0
decorator==5.1.1
deepdiff==5.8.1
defusedxml==0.7.1
dill==0.3.4
efficientnet-pytorch==0.7.1
entrypoints==0.4
evaluate==0.4.0
executing==0.8.3
experiment-impact-tracker==0.1.8
fastjsonschema==2.15.3
filelock==3.6.0
Flask==2.2.2
fonttools==4.34.4
frozenlist==1.3.0
fsspec==2022.3.0
future==0.18.2
geocoder==1.38.1
geographiclib==1.52
geopy==2.2.0
google-auth==2.9.1
google-auth-oauthlib==0.4.6
grpcio==1.48.0
huggingface-hub==0.12.0
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
json5==0.9.6
jsonschema==4.5.0
jupyter-client==7.3.0
jupyter-core==4.10.0
jupyter-server==1.17.0
jupyterlab==3.4.0
jupyterlab-pygments==0.2.2
jupyterlab-server==2.13.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==6.0.2
multiprocess==0.70.12.2
nbclassic==0.3.7
nbclient==0.6.2
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==3.0
notebook==6.4.11
notebook-shim==0.1.0
numpy==1.21.5
oauthlib==3.2.0
opencv-python==4.6.0.66
ordered-set==4.1.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
plotly==5.16.1
progiter==1.0.1
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==8.0.0
py4j==0.10.9.5
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
PyLaTeX==1.4.1
pynvml==11.4.1
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
ratelim==0.1.6
regex==2022.4.24
responses==0.18.0
retrying==1.3.4
rsa==4.9
s3transfer==0.6.0
sacremoses==0.0.53
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
sentencepiece==0.1.97
Shapely==1.8.5
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.2.0
tenacity==8.2.3
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.13.3
thop==0.1.1.post2207130030
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.12.1
torch==1.12.0
torch-tb-profiler==0.4.3
torchaudio==0.12.0
torchvision==0.13.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
transformers @ git+https://github.com/huggingface/transformers@ea55bd86b9a452c87c5383afc707ab7d710a3043
ujson==5.5.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.2.2
xxhash==3.0.0
yarl==1.7.2
